Moved cheevos files to their own folder in preparation for the changes to support leaderboards
