+++ managers/cheat_manager.c
@@ -30,7 +30,7 @@
 #endif
 
 #ifdef HAVE_CHEEVOS
-#include "../cheevos.h"
+#include "../cheevos/cheevos.h"
 #endif
 
 #include "cheat_manager.h"
