+++ menu/menu_displaylist.c
@@ -37,7 +37,7 @@
 #endif
 
 #ifdef HAVE_CHEEVOS
-#include "../cheevos.h"
+#include "../cheevos/cheevos.h"
 #endif
 
 #ifdef HAVE_NETWORKING
