+++ network/httpserver/httpserver.c
@@ -29,7 +29,7 @@
 #include "../../core.h"
 #include "../../gfx/video_driver.h"
 #include "../../managers/core_option_manager.h"
-#include "../../cheevos.h"
+#include "../../cheevos/cheevos.h"
 #include "../../content.h"
 
 #define BASIC_INFO "info"
